--- core/core-api/src/main/java/com/textadventure/core/api/ActionDao.java
@@ -1,11 +1,0 @@
-package com.textadventure.core.api;
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public interface ActionDao {
-    ActionEntity getActionById(int id);
-    List<ActionEntity> getActionsByNode(NodeEntity node);
-    void saveAction(ActionEntity action);
-    void deleteAction(ActionEntity action);
-    void updateAction(ActionEntity action);
-}
