+++ quest-editor/nctask/src/main/java/com/nc/example/DBCarSaver.java
@@ -1,0 +1,7 @@
+package com.nc.example;
+
+/**
+ * Created by dimko_000 on 15.12.2015.
+ */
+public class DBCarSaver {
+}
