+++ quest-editor/nctask/src/main/java/com/nc/example/Evacuator.java
@@ -1,0 +1,10 @@
+package com.nc.example;
+
+/**
+ * Created by dimko_000 on 14.12.2015.
+ */
+public class Evacuator implements Runnable {
+    public void run() {
+
+    }
+}
