+++ quest-editor/nctask/src/main/java/com/nc/example/Master.java
@@ -1,0 +1,7 @@
+package com.nc.example;
+
+/**
+ * Created by dimko_000 on 14.12.2015.
+ */
+public class Master {
+}
