--- quest-editor/quest-editor-api/.gitignore
@@ -1,15 +1,0 @@
-*.class
-*.iml
-*.idea
-target/
-
-# Mobile Tools for Java (J2ME)
-.mtj.tmp/
-
-# Package Files #
-*.jar
-*.war
-*.ear
-
-# virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
-hs_err_pid*
