--- quest-editor/quest-editor-api/src/main/java/org/textadventure/questeditor/api/Condition.java
@@ -1,8 +1,0 @@
-package org.textadventure.questeditor.api;
-
-/**
- * Created by dimko_000 on 06.12.2015.
- */
-public interface Condition {
-    boolean checkCondition();
-}
