--- quest-editor/quest-editor-api/src/main/java/org/textadventure/questeditor/api/Stat.java
@@ -1,11 +1,0 @@
-package org.textadventure.questeditor.api;
-
-/**
- * Created by dimko_000 on 06.12.2015.
- */
-public interface Stat {
-    void setName(String name);
-    String getName();
-    void setValue(double value);
-    double getValue();
-}
