+++ quest-editor/quest-editor-impl/src/main/java/org/my/adventure/questeditor/impl/ActionController.java
@@ -1,0 +1,9 @@
+package org.my.adventure.questeditor.impl;
+import or
+/**
+ * Created by dimko_000 on 10.12.2015.
+ */
+public class ActionController {
+     actionDao;
+
+}
