+++ quest-editor/quest-editor-impl/src/main/java/org/my/adventure/questeditor/impl/ConditionController.java
@@ -1,0 +1,9 @@
+package org.my.adventure.questeditor.impl;
+
+
+/**
+ * Created by dimko_000 on 10.12.2015.
+ */
+public class ConditionController {
+   // ConditionDao conditionDao;
+}
