+++ quest-editor/quest-editor-impl/src/main/java/org/my/adventure/questeditor/impl/QuestController.java
@@ -1,0 +1,8 @@
+package org.my.adventure.questeditor.impl;
+
+/**
+ * Created by dimko_000 on 10.12.2015.
+ */
+public class QuestController {
+   // QuestDao questDao;
+}
