--- quest-editor/quest-editor-impl/src/main/java/org/textadventure/questeditor/impl/ActionController.java
@@ -1,11 +1,0 @@
-package org.textadventure.questeditor.impl;
-
-import com.textadventure.core.api.ActionDao;
-
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public class ActionController {
-    ActionDao actionDao;
-
-}
