--- quest-editor/quest-editor-impl/src/main/java/org/textadventure/questeditor/impl/ConditionController.java
@@ -1,10 +1,0 @@
-package org.textadventure.questeditor.impl;
-
-import com.textadventure.core.api.ConditionDao;
-
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public class ConditionController {
-    ConditionDao conditionDao;
-}
