--- quest-editor/quest-editor-impl/src/main/java/org/textadventure/questeditor/impl/NodeController.java
@@ -1,10 +1,0 @@
-package org.textadventure.questeditor.impl;
-
-import com.textadventure.core.api.NodeDao;
-
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public class NodeController {
-    NodeDao nodeDao;
-}
