--- quest-editor/quest-editor-impl/src/main/java/org/textadventure/questeditor/impl/QuestController.java
@@ -1,10 +1,0 @@
-package org.textadventure.questeditor.impl;
-
-import com.textadventure.core.api.QuestDao;
-
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public class QuestController {
-    QuestDao questDao;
-}
