--- quest-editor/quest-editor-impl/src/main/java/org/textadventure/questeditor/impl/TransitionController.java
@@ -1,11 +1,0 @@
-package org.textadventure.questeditor.impl;
-
-import com.textadventure.core.api.TransitionDao;
-
-/**
- * Created by dimko_000 on 10.12.2015.
- */
-public class TransitionController {
-    TransitionDao transitionDao;
-
-}
