+++ text-adventure-app/text-adventure-app-web/src/main/webapp/editor.xhtml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN"
+        "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml"
+      xmlns:f="http://java.sun.com/jsf/core"
+      xmlns:h="http://java.sun.com/jsf/html">
+
+<h:head>
+    <title>Hello World!</title>
+</h:head>
+<h:body>
+    <h3>#{test.getHelloWorld()}</h3>
+</h:body>
+</html>
